Introduce `MixteraDataCollection` (#5)
This PR introduces the `MixteraDataCollection` in its local variant,
allowing to register datasets. The next step will be to create an
inverted index on some properties as done by @xzyaoi in his MVP, and to
add new properties to the dataset by running pre-query operators. There
is one example for how we could build an inverted index, hardcoded to
SlimPajama.