Integrate docstrings with Sphinx (#23)
This PR creates automatic documents for Data Collections into sphinx
docs. Related Issue: #14.

The main change is about the `__init__.py`, which needs to explicitly
export classes that we want users to use. I now manually set the main
classes to be public and removed the exported files - this should be
enough for the test cases and also the auto docs. Let me know if you'd
like the export the files as well :)

To preview the documents, `cd` into the `docs` folder and run `make
watch`.