Add ability to extend index (#16)
This PR adds the ability to expand the existing index based on new
attributes, e.g., classification result.

This is implemented using a local executor, but reflects the Ray
programming pattern, such that implementing it using Ray will be
straightforward. Currently, the result is not persisted, as the index is
not persisted to DB yet.