Feature: Persist indices to sqlite (#28)
With `add_property` function, the new property and the indices will be
added to the database for future usage.

The main changes:
- Now there's no `hacky_index`, instead it is just called index for now.
- New function `get_index(property_name: Optional[str])` to access the
index. Optionally you can give a property_name that you're interested
in. If not specified, all properties will be returned.
- A new utility function to convert numpy objects into native python
objects (so we can easily insert into sqlite).
- When register dataset or add_property, the _index in the memory won't be updated. Instead it will just be written into the database.
- `get_index` now checks if the property exists in the in_memory _index, if so it won't load from the database.