Implement reading samples from JSONLDataset (#27)
For a given list of ranges per file, this PR enables to get the sample
data for these ranges as an iterator.

@xzyaoi: The idea I have in my head is that the result of the operator
graph evaluation is a list per dataset, which, per file in the dataset,
contains ranges. Hence, I will provide in another PR a function for the
LocalCollection that calls this function on the dataset.