Read samples from ranges from `DataCollection` (#29)
Given a list of ranges per file per dataset, this implements an iterator
over all the samples from this file. This extends PR #27 (so #27 needs
to be reveiwed and merged first).

The idea is to store a function that given some basic unit (depends on
the dataset, in jsonl, it is one line) parses the actual sample data. We
can use this UDF then while iterating.