Make `LocalPropertyCalculationExecutor` decoupled from jsonl format (#30)
This adds a check whether we are actually reading JSONLDatasets in the
`LocalPropertyCalculationExecutor` and fails otherwise. We cannot
implement this on the `JSONLDataset` class directly because each
executor has different requirements - Ray will need some other
functionality depending on dataset type than local. Hence, it is done
directly in the executor.

Needs #29 to be merged first (then change base to main and fix potential
conflicts).

Solves #19.