Removed the use of the legacy IndexType (#47)
Removed the use of the legacy `IndexType` and `UncompressedIndexType`
formerly defined in `core/datacollection/__init__.py`. Also updated the
returned type of `build_file_index` in `CroissantDataset` to `None`
(such that it matches the `Dataset` interface), and the return type of
`get_index` to `Optional[InMemoryDictionaryRangeIndex]` in the
`MixteraDataCollection` such that it matches the current
`LocalDataCollection` implementation.