Use `mp.Lock` instead of `threading.Lock` to make examples run in `spawn` context (#68)
On macOS, the multiprocessing context is `spawn` instead of `fork`. In
the spawn context, we cannot pickle a threading.Lock. Hence, we use a
mp.Lock instead. While this is a bit more expensive, it's not on the hot
path (only on query registration), and hence should be ok. This allows
us to run the examples on macOS again.

Actually, I am not sure whether and why this worked before I left.