Chunking (#66)
Adds query result chunking, based either on a chunk size without a
Mixture object or with a Mixture object that specifies the mass of
various property combinations within a chunk. The current chunking
mechanism also ensures that property duplicates (i.e. instances that
have several relevant values for a given property) are deduplicated. In
other words, no instance will be seen more than once during an epoch and
it will be assigned only one value per property (even if more are
viable; e.g. a text that is both English and German will only be
classified as one). The generated chunks will have the following
structure (as given by the `ChunkerIndex` type):

```
{
  "prop_0_name:prop_0_value;prop_1_name:prop_1_value...": {
    dataset_0_id: {
      file_0_id: [
        (lo_bound_0, hi_bound_0), 
        ...
      ],
      ...
    },
    ... 
  },
  ...
}
```

Concretely, an example might look like: 

```
{
  "language:french;topic:law": {
    0: {
      0: [(0, 6)]
    }
  }, 
  "language:english;topic:medicine": {
    0: {
      0: [(50, 54)]
    }
  }
}
```

for a mixture for the form: 

```
{
  "language:french;topic:law": 0.6,
  "language:english;topic:medicine": 0.4,
}
```

given a `chunk_size=10`.

When using a mixture, it should be noted that the current logic produces
as many chunks as it possibly can before running out of any one
component for the chunks. For instance, if for a chunk of size 100, with
equal distribution among two components (i.e. 50 each per chunk), one
with 1000 base elements, while the other with 500 base elements, it
follows that we can only produce `10 = min(1000 / 50, 500 / 50)` chunks.
This is the intended behavior as otherwise the mixture would be
consistently violated.

Known limitations:

- When only supplying a `chunk_size` (i.e. without a mixture), then
chunking is applied on a per-property basis without any cross-property
chunking. This means that the final chunk within a property might have a
lower cardinality than `chunk_size`. This should not be an issue in
practice, though since instances are seen exactly once regardless.
- When chunks are composed using mixtures, the dataset cannot always
guarantee that the final batch will contain the desired amount from each
mixture. For this final batch it follows that the mixture might be
violated due to dataset limitations.