Fix the query result pickling to work on more platforms (#78)
Linux is using fork for creating new processes which does not require
pickling the QueryResult while macOS is using spawn. This resulted in
different behaviour of the CI pipeline (integrationtest of pyTorch
dataset) on the two platforms. To make the behaviour more robust we had
to change the pickling of QueryResult to on one hand do not pickle the
generator and secondly pickle more of the state using dill.

Co-authored-by:
MaxiBoether <2116466+MaxiBoether@users.noreply.github.com>