Introduce `ChunkDistributor` to correctly distribute chunks across data parallel nodes (#77)
This PR introduces the `ChunkDistributor`. This layer is between the
workers requesting chunks and the QueryResult. It ensures (a) different
workers on a machine get different chunks, (b) workers with the same ID
within the same DP group get the same chunks in the same order (e.g.,
worker 0 on node 0 in dp group 0 gets the same as worker 0 on node 1 in
dp group 0, and worker 0 on node 0 in dp group 1 gets the same as worker
0 on node 1 in dp group 1, but node the same between those two dp
groups). This allows us to use Mixtera in Nanotron's 3D parallel
training.