Implement `InferringMixture` as an alternative to `ArbitraryMixture` (#96)
The `ArbitraryMixture` does not give us any guarantees: It can yield
chunks of _arbitrary_ content with spurious correlations. The
`InferringMixture` addresses this issue (and is probably better for
convergence): We infer the distribution of properties in the query
result, and yield chunks that follow the same distribution as the query
result.