Cache `Query` to avoid re-inverting index if nothing changes (#102)
This PR implements a simple caching policy based on the string
representation of a query. It pickles the query object and restores the
query result when a matching query is found. This avoids re-inverting
the index for queries that share the same static part (mixture can be
different).

This PR also fixes some issues in the huggingface dataset in case we run
data parallel training with num_workers = 0.

Closes #100.