Allow using parallel reading in dataset workers (#106)
The file reading mechanism introduced in #67 uses multiple processes to
read files in parallel. However, when having more than 0 torch data
loading workers, this is a problem. The workers are spawned as demonic
processes, and those in Python cannot have children, i.e., we get an
exception as soon as the reading processes are spawned.

This PR fixes this issue by hacking around this limitation and
implementing proper cleanup by checking whether the parent died. Also we
change the test that should have caught this earlier to actually catch
it.

Solves #103.