feat: Add `HierarchicalStaticMixture`  (#115)
In the pr, the issue https://github.com/eth-easl/mixtera/issues/104 is
addressed so that users can define mixtures in a more hiearchical
manner. The new class is almost identical to StaticMixtures except the
input format is more intuitive. To define the HiearchicalStaticMixture
the following input syntax is designed as shown in the test file:
`{
"language": "JavaScript",
                    "portion": 0.5,
                    "submixture": [
                        {
                            "licence": "CC",
                            "portion": 0.8,
                            "submixture": []
                        },
                        {
                            "licence": "All rights reserved.",
                            "portion": 0.2,
                            "submixture": []
                        }
                    ],
},`

---------

Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>
Co-authored-by: Beste Gueney <bgueney@sgs-gpu03.ethz.ch>