feat: WebDataset Support (#136)
This PR implements support for the WebDataset format, using the `wids`
library for IO operations. To stay consistent with current abstractions
in Mixtera, a file will refer to a single .tar file which contains
multiple samples and the `line_number` refers to the index of the sample
within that .tar file as the samples are sorted and numbered in the
WebDataset format.

There is still a problem with how the server_connection and our
FileSystems abstraction works, need to discuss how to work around it.