Creating the Mixture Schedule (#139)
In this pr, the mixture schedule is defined. With this definition, it
will be possible to define a training schedule and retrieve the correct
mixture to train on by looking at the training steps taken from the
feedback defined here https://github.com/eth-easl/mixtera/pull/133. In
the upcoming pr's, the integration of mixture schedule and client
feedback will be implemented.

---------

Co-authored-by: Maximilian Böther <2116466+MaxiBoether@users.noreply.github.com>