feat: Simplify chunk generation a bit (#175)
While writing up the chunk generation algorithm, I realized we -- I
think -- do not need a progress per key flag. In my understanding, if we
did not obtain enough samples from a component iterator after a single
pass through the component iterators, a second pass will not help, since
all iterators are depleted.

This is only true if the iterators not "just for fun" yield too little
data. What is necessary is that they either yield n samples, or < n and
then they are empty but the next try will yield 0.