feat: Modernize packaging (#184)
This PR moves away from mamba/conda to a pip-based setup. We also use
towards a `pyproject.toml` setup to clean up our configurations. The
CMake setup is also made more robust, and we migrate the CI to pixi+uv
since it is the fastest way to set up a new environment.